requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.6
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
colorthief==0.2.1
discord.py==1.7.3
emoji==1.2.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
Pillow==8.2.0
pylint==2.8.3
requests==2.25.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wrapt==1.12.1
yarl==1.6.3
